+++ R/help.R
@@ -2,11 +2,11 @@
 #' @importFrom shiny textOutput
 helpUI <- function(id, tab) {
     ns <- NS(id)
-    
+
     # if (requireNamespace("parallel", quietly = TRUE)) {
     #     cores <- parallel::detectCores()
     #     coresInput <- tagList(
-    #         sliderInput(ns("cores"), h4("Number of cores"), value=1, min=1, 
+    #         sliderInput(ns("cores"), h4("Number of cores"), value=1, min=1,
     #                     step=1, max=cores, width="auto", post=" core(s)"),
     #         helpText("A total of", cores, "cores were detected.")
     #     )
@@ -14,17 +14,17 @@ helpUI <- function(id, tab) {
     #     coresInput <- numericInput(ns("cores"), h4("Number of cores"), value=1,
     #                                min=1, step=1, width="auto")
     # }
-    
+
     link <- function(href, ..., target="_blank") {
         a <- tags$a(href=href, target=target, ...)
     }
-    
+
     linkItem <- function(href, ..., target="_blank") {
         a <- link(href, ..., target=target)
         a <- tags$li(class="list-group-item", a)
         return(a)
     }
-    
+
     guiLink <- paste0("https://nuno-agostinho.github.io/psichomics/",
                       "articles/GUI_tutorial.html")
     cliLink <- paste0("https://nuno-agostinho.github.io/psichomics/",
@@ -35,7 +35,7 @@ helpUI <- function(id, tab) {
                               "articles/AS_events_preparation.html")
     tutorials <- div(
         class="panel", class="panel-default",
-        div(class="panel-heading", icon("file-text"), tags$b("Tutorials")),
+        div(class="panel-heading", icon("file-alt"), tags$b("Tutorials")),
         tags$ul(
             class="list-group",
             linkItem(guiLink, "Visual interface tutorial"),
@@ -44,7 +44,7 @@ helpUI <- function(id, tab) {
                      "Loading SRA and user-provided RNA-seq data"),
             linkItem(customAnnotLink,
                      "Preparing custom alternative splicing annotations")))
-    
+
     supportLink  <- "https://support.bioconductor.org/t/psichomics/"
     githubIssues <- "https://github.com/nuno-agostinho/psichomics/issues/new"
     feedback <- div(
@@ -53,14 +53,14 @@ helpUI <- function(id, tab) {
         tags$ul(class="list-group",
                 linkItem(supportLink, "Questions and general support"),
                 linkItem(githubIssues, "Suggestions and bug reports")))
-    
+
     groupSite <- "http://imm.medicina.ulisboa.pt/group/distrans/"
     immSite   <- "http://imm.medicina.ulisboa.pt"
     about <- div(
         class="panel", class="panel-default",
         div(class="panel-heading",
             icon("info-circle"), tags$b("About")),
-        div(class="panel-body", 
+        div(class="panel-body",
             "psichomics is an interactive R package for integrative analyses",
             "of alternative splicing and gene expression from large",
             "transcriptomic datasets including those from",
@@ -72,9 +72,9 @@ helpUI <- function(id, tab) {
             "project, as well as user-provided data."),
         tags$ul(class="list-group",
                 tags$li(class="list-group-item",
-                        tags$b("Developer:"), 
+                        tags$b("Developer:"),
                         link("mailto:nunodanielagostinho@gmail.com",
-                             "Nuno Saraiva-Agostinho", icon("envelope-o"))),
+                             "Nuno Saraiva-Agostinho", icon("envelope"))),
                 tags$li(class="list-group-item",
                         tags$b("Supervisor:"), "Nuno L. Barbosa-Morais"),
                 linkItem(groupSite, "Disease Transcriptomics lab"),
@@ -82,7 +82,7 @@ helpUI <- function(id, tab) {
                 tags$li(class="list-group-item",
                         tags$small(class="help-block", style="margin: 0;",
                                    style="text-align: right;", "2015-2020"))))
-    
+
     credits <- c("Lina Gallego", "Marie Bordone", "Mariana Ferreira",
                  "Teresa Maia", "Carolina Leote", "Juan Carlos Verjan",
                  "Bernardo de Almeida")
@@ -95,7 +95,7 @@ helpUI <- function(id, tab) {
             "This work would not be possible without the support of current",
             "and former members of Nuno Morais lab. Thank you for your help."),
         do.call(tags$ul, c(credits, list(class="list-group"))))
-    
+
     tab(title="Help", icon="question", linkToArticles(),
         h2("Settings", style="margin-top: 0;"), fluidRow(
             # column(4, coresInput),
@@ -128,11 +128,11 @@ helpServer <- function(input, output, session) {
                   formatC(283.5837243243,
                           digits=getPrecision(), format="f")))
     })
-    
+
     observe({
         setSignificant(input$significant)
         output$significantExample <- renderText(
-            paste("Example:", 
+            paste("Example:",
                   formatC(5.849371935e-06,
                           getSignificant(), format="g")))
     })